A few stuff removal of the ZK
