+++ src/main/resources/kafka_versions.json
@@ -1,7 +1,5 @@
 {
   "kafkaVersions": {
-    "3.8.0": "quay.io/strimzi-test-container/test-container:latest-kafka-3.8.0",
-    "3.8.1": "quay.io/strimzi-test-container/test-container:latest-kafka-3.8.1",
     "3.9.0": "quay.io/strimzi-test-container/test-container:latest-kafka-3.9.0",
     "4.0.0": "quay.io/strimzi-test-container/test-container:latest-kafka-4.0.0"
   }
