+++ src/test/java/io/strimzi/test/container/StrimziConnectClusterIT.java
@@ -107,7 +107,7 @@ public class StrimziConnectClusterIT {
 
     @Test
     public void testKafkaVersion() throws Exception {
-        String version = "3.8.1";
+        String version = "3.9.0";
         connectCluster = new StrimziConnectCluster.StrimziConnectClusterBuilder()
                 .withGroupId("my-cluster")
                 .withKafkaCluster(kafkaCluster)
